backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
numpy>=1.24.0
httpx>=0.24.0
